SystemSpoofer, capture mode, CPUID key fix, proxy improvements
New features:
- SystemSpoofer: INT3-based interception of SGDT/SIDT/SLDT/STR/XGETBV/RDMSR
  via VirtualQuery address walk with MEM_IMAGE exclusion (prevents self-corruption)
- SyscallDispatch: centralized syscall dispatch layer
- CaptureLogger + capture mode: logs all fingerprinting queries to disk
  without spoofing (enabled via config/capture.ini)
- capture_tool + msr_reader: new tools for fingerprint collection

Fixes:
- CPUID hash key: uint32_t (leaf << 8 | subleaf) collided for extended leaves
  0x80000000+ vs 0x0 / 0x1. Fixed with uint64_t ((uint64_t)leaf << 32 | subleaf)
- PEB spoof: removed ProcessHeap.Flags/ForceFlags writes (offsets differ per
  Windows version, corrupt heap metadata on Win10 22H2+)
- Launcher event name: SymbioteEngineReady -> Symbiote_EngineReady
- Launcher timeout: restored to 5000ms

WHP improvements:
- VcpuManager: trampoline restore on destruction, ReadVcpuRegs/WriteVcpuRegs,
  syscall dispatch init
- Additional CPUID leaves: 0x07 subleaf 1, 0x14 subleaf 1, 0x19, 0x1F

Proxy DLLs:
- crypt32_proxy: CryptGetProvParam spoofed container name
- kernel32_proxy: GetVolumeInformationW, GetWindowsDirectoryW hooks
- kernelbase_proxy, ntdll_proxy: updates

Config:
- Feature per-section toggle format [feature].status (backward compat with [spoof])
- system_spoofer section for GDT/IDT/XGETBV override values
- capture config section

+++ src/engine/kernel/SystemProfile.h
@@ -56,7 +56,7 @@ public:
     void LoadAmdRyzen9_5950X();
 
 private:
-    std::unordered_map<uint32_t, CpuidLeaf> m_cpuidLeaves;
+    std::unordered_map<uint64_t, CpuidLeaf> m_cpuidLeaves;
     uint64_t m_tscFrequency;
     uint64_t m_tscOffset;
     uint64_t m_apicFrequency;
